Remove ./cdrom, add scripts/ partman/ and qubes/
Laurent Boulard had personal informations in ./cdrom such as its public ssh
keys.

I consider that the repository should restraint to include such informations and
provide now a mechanism to include files in private/ that will be exposed in the
iso as /cdrom/private/.

The same mechanism is used to include partman/ (partman-auto partitioning
recipes) and scripts/ (sh scripts to be executed).

The qubes/preseed.cfg is my testing preseed configuration.